% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{ufp_aggregate}
\alias{ufp_aggregate}
\title{Aggregate PUFP data}
\usage{
ufp_aggregate(
  df,
  unit = "5 seconds",
  floor_or_celiling = "floor",
  summary_fun = "median"
)
}
\arguments{
\item{df}{an object created by \code{ufp_read()} or \code{ufp_batch_read()}.}

\item{unit}{a character string specifying a time unit or a multiple of a
unit to be rounded}

\item{floor_or_celiling}{either 'floor'(\code{\link[lubridate]{floor_date}})
or 'ceiling'(\code{\link[lubridate]{ceiling_date}}). Default = 'floor.'}

\item{summary_fun}{summary function (i.e. 'mean', 'median').
Default = 'median.'}
}
\value{
a tibble
}
\description{
\code{ufp_aggregate()} condenses PUFP time series data to a specified time unit.
}
\examples{
\dontrun{

ufp_aggregate(df,
  unit = "5 seconds", floor_or_ceiling = "floor",
  summary_fun = "median"
)
}
}
