% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_read.R
\name{ufp_batch_read}
\alias{ufp_batch_read}
\title{Read multiple PUFP .txt files}
\usage{
ufp_batch_read(
  paths,
  event_threshold = 10,
  tz = "America/New_York",
  truncate_ufp = TRUE,
  coords = TRUE,
  ufp_check = FALSE,
  participant_id = NULL,
  sample_col = NULL,
  sample_id = NULL
)
}
\arguments{
\item{paths}{a vector of paths}

\item{event_threshold}{numeric; difftime value in minutes to differentiate
sampling events. Consecutive measurements that exceed this threshold are separated
into new sampling events. Default = 10.}

\item{tz}{a character string that specifies which time zone to parse the
date with. Default = 'America/New_York.'}

\item{truncate_ufp}{logical; truncate UFP concentration? If TRUE
(the default), UFP concentrations above 250K will be right censored.}

\item{coords}{logical; arse GPS string to derive latitude and longitude?
Default = TRUE.}

\item{ufp_check}{check for invalid UFP measurements.  If TRUE, new columns
named \code{UFP_NA} and \code{UFP_Invalid} are created to flag missing and potentially
invalid UFP concentrations.
#' @param participant_id  user defined string to denote a personal identifier.
This is useful if the PUFP is deployed during personal sampling.  If specified,
a new column is created ('ID'). Default is NULL.}

\item{sample_col}{user defined character string specifying the name of the
column to denote sample ID. Default is NULL.}

\item{sample_id}{user defined string to denote sample ID. If assigned, a
value must also be supplied to \code{sample_col}. Default is NULL.}
}
\value{
a tibble. Additional columns are created for 'Sampling_Event' and
'Sampling_Day.'  Sampling events represent discreet intervals in sampling
defined by the \code{event_threshold}.  Sampling day is derived from the date
of the measurements.  Multiple sampling events may occur within a sampling
day.
}
\description{
\code{ufp_batch_read()} imports and cleans multiple PUFP text files as a single
data frame.  Measurements from different monitoring sessions are denoted by
the 'Sampling_Event' column, which is defined by \code{event_threshold}.
}
\examples{
\dontrun{

ufp_batch_read(c(path1, path2, path3),
  event_threshold = 10,
  tz = "America/New_York", truncate_ufp = TRUE, coords = TRUE, ufp_check = FALSE,
  participant_id = NULL, sample_col = NULL, sample_id = NULL
)
}
}
