% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularize.R
\name{ufp_circularize}
\alias{ufp_circularize}
\title{Title}
\usage{
ufp_circularize(df, circvar_threshold = 0.7, window = 60,
  cluster_threshold = NULL)
}
\arguments{
\item{df}{}

\item{circvar_threshold}{}

\item{window}{}

\item{cluster_threshold}{}
}
\description{
Title
}
