% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_read.R
\name{ufp_batch_read}
\alias{ufp_batch_read}
\title{Read multiple PUFP .txt files}
\usage{
ufp_batch_read(
  paths,
  event_threshold = 10,
  tz = "America/New_York",
  truncate_ufp = TRUE,
  coords = TRUE
)
}
\arguments{
\item{paths}{a vector of paths}

\item{event_threshold}{numeric; difftime value in minutes to differentiate
sampling events. Consecutive measurements that exceed this threshold are separated
into new sampling events. Default = 10.}

\item{tz}{a character string that specifies which time zone to parse the
date with. Default = 'America/New_York.'}

\item{truncate_ufp}{logical; truncate UFP concentration? If TRUE
(the default), UFP concentrations above 250K will be right censored.}

\item{coords}{logical; arse GPS string to derive latitude and longitude?
Default = TRUE.}
}
\value{
a tibble. Additional columns are created for 'Sampling_Event' and
'Sampling_Day.'  Sampling events represent discreet intervals in sampling
defined by the \code{event_threshold}.  Sampling day is derived from the date
of the measurements.  Multiple sampling events may occur within a sampling
day.
}
\description{
\code{ufp_batch_read()} imports and cleans multiple PUFP text files as a single
data frame.  Measurements from different monitoring sessions are denoted by
the 'Sampling_Event' column, which is defined by \code{event_threshold}.
}
\examples{
\dontrun{

ufp_batch_read(c(path1, path2, path3),
  event_threshold = 10,
  tz = "America/New_York", truncate_ufp = TRUE, coords = TRUE
)
}
}
