% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularize.R
\name{ufp_circularize}
\alias{ufp_circularize}
\title{Title}
\usage{
ufp_circularize(
  df,
  circvar_threshold = 0.7,
  window = 60,
  cluster_threshold = NULL,
  show_circvar = FALSE
)
}
\arguments{
\item{df}{NEED DEF}

\item{circvar_threshold}{NEED DEF}

\item{window}{NEED DEF}

\item{cluster_threshold}{NEED DEF}

\item{show_circvar}{NEED DEF}
}
\value{
NEED DEF
}
\description{
Title
}
\examples{
\dontrun{

ufp_circularize(df, circvar_threshold = .7, window = 60, cluster_threshold = NULL, 
show_circvar = FALSE)
}
}
