% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularize.R
\name{ufp_circularize}
\alias{ufp_circularize}
\title{Title}
\usage{
ufp_circularize(
  df,
  circvar_threshold = 0.7,
  window = 60,
  cluster_threshold = NULL,
  show_circvar = FALSE
)
}
\arguments{
\item{df}{}

\item{circvar_threshold}{}

\item{window}{}

\item{cluster_threshold}{}
}
\value{

}
\description{
Title
}
\examples{
#' \dontrun{

ufp_circularize(df, circvar_threshold = .7, window = 60, cluster_threshold = NULL, show_circvar = FALSE)
}
}
