% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{ufp_move}
\alias{ufp_move}
\title{Calculate Speed and Azimuth of PUFP Measurements}
\usage{
ufp_move(df)
}
\arguments{
\item{df}{data frame created by `ufp_read() `or `ufp_batch_read()`}
}
\value{
a data frame.  Speed is expressed in meters per second; azimuth is 
expressed in degrees.
}
\description{
Calculates the speed and azimuth (bearing) of consecutive PUFP measurements.
If multiple sampling events are included in the input file, 
speed and azimuth are calcuated by event.
}
\examples{
\dontrun{

ufp_move(df)
}
}
