% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ufp_read}
\alias{ufp_read}
\title{Read PUFP .txt File}
\usage{
ufp_read(path, tz = "America/New_York", truncate_ufp = TRUE,
  coords = TRUE)
}
\arguments{
\item{path}{a path.}

\item{tz}{a character string that specifies which time zone to parse the 
date with. Default = "America/New_York."}

\item{truncate_ufp}{truncate UFP concentration? If TRUE (the default), UFP
concentrations above 250K will be right censored.}

\item{coords}{parse GPS string to derive latitude and longitude? 
Default = TRUE.}
}
\value{
a tibble.
}
\description{
Reads and cleans text file output from PUFP sensor.
}
\examples{
\dontrun{

ufp_read(path, tz = "America/New_York", truncate_ufp = TRUE, coords = TRUE)
}
}
